# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YU-GI-OH! Api Rest",
        "description": "Implementation of Yu-Gi-Oh! Api Rest",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local enviroment"
        }
    ],
    "paths": {
        "/api/cards": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "List all cards.",
                "operationId": "card-index",
                "responses": {
                    "200": {
                        "description": "Get a cards collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cards"
                ],
                "summary": "Create a card.",
                "operationId": "card-store",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "first_edition",
                                    "serial_code",
                                    "type",
                                    "amount",
                                    "subtype_id",
                                    "image_id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Card's ID",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "name": {
                                        "description": "Card's name",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "description": {
                                        "description": "Card's description",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "first_edition": {
                                        "description": "Is first edition",
                                        "type": "boolean",
                                        "format": ""
                                    },
                                    "serial_code": {
                                        "description": "Card's serial code",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "type": {
                                        "description": "Card's type",
                                        "type": "string",
                                        "format": "",
                                        "enum": [
                                            "Monster",
                                            "Magic",
                                            "Trap"
                                        ]
                                    },
                                    "attack": {
                                        "description": "Attack of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "defense": {
                                        "description": "Defense of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "star": {
                                        "description": "Quality of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "amount": {
                                        "description": "Amount of card",
                                        "type": "number",
                                        "format": ""
                                    },
                                    "subtype_id": {
                                        "description": "Card's subtype",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "image_id": {
                                        "description": "Card's image",
                                        "type": "integer",
                                        "format": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/cards/{id}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Get a specific card.",
                "operationId": "card-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card obteined.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Cards"
                ],
                "summary": "Update a specific card.",
                "operationId": "card-update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Card's ID",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "name": {
                                        "description": "Card's name",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "description": {
                                        "description": "Card's description",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "first_edition": {
                                        "description": "Is first edition",
                                        "type": "boolean",
                                        "format": ""
                                    },
                                    "serial_code": {
                                        "description": "Card's serial code",
                                        "type": "string",
                                        "format": ""
                                    },
                                    "type": {
                                        "description": "Card's type",
                                        "type": "string",
                                        "format": "",
                                        "enum": [
                                            "Monster",
                                            "Magic",
                                            "Trap"
                                        ]
                                    },
                                    "attack": {
                                        "description": "Attack of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "defense": {
                                        "description": "Defense of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "star": {
                                        "description": "Quality of card",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "amount": {
                                        "description": "Amount of card",
                                        "type": "number",
                                        "format": ""
                                    },
                                    "subtype_id": {
                                        "description": "Card's subtype",
                                        "type": "integer",
                                        "format": ""
                                    },
                                    "image_id": {
                                        "description": "Card's image",
                                        "type": "integer",
                                        "format": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Card updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/card/{id}": {
            "delete": {
                "tags": [
                    "Cards"
                ],
                "summary": "Delete a card.",
                "operationId": "card-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card deleted."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CardResource": {
                "title": "Card",
                "description": "Card information",
                "properties": {
                    "id": {
                        "description": "Card's ID",
                        "type": "integer",
                        "format": ""
                    },
                    "name": {
                        "description": "Card's name",
                        "type": "string",
                        "format": ""
                    },
                    "description": {
                        "description": "Card's description",
                        "type": "string",
                        "format": ""
                    },
                    "first_edition": {
                        "description": "Is first edition",
                        "type": "boolean",
                        "format": ""
                    },
                    "serial_code": {
                        "description": "Card's serial code",
                        "type": "string",
                        "format": ""
                    },
                    "type": {
                        "description": "Card's type",
                        "type": "string",
                        "format": "",
                        "enum": [
                            "Monster",
                            "Magic",
                            "Trap"
                        ]
                    },
                    "attack": {
                        "description": "Attack of card",
                        "type": "integer",
                        "format": ""
                    },
                    "defense": {
                        "description": "Defense of card",
                        "type": "integer",
                        "format": ""
                    },
                    "star": {
                        "description": "Quality of card",
                        "type": "integer",
                        "format": ""
                    },
                    "amount": {
                        "description": "Amount of card",
                        "type": "number",
                        "format": ""
                    },
                    "subtype": {
                        "$ref": "#/components/schemas/SubtypeResource"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Dark Magician",
                    "description": "Desciption of Dark Magician",
                    "first_edition": true,
                    "serial_code": 190124914710293,
                    "type": "Monster",
                    "attack": 2500,
                    "defense": 2000,
                    "star": 9,
                    "amount": 1500,
                    "subtype": {
                        "id": 1,
                        "name": "Normal Monster"
                    },
                    "image_id": {
                        "id": 1,
                        "name": "Dark Margician",
                        "extension": "jpg",
                        "file": "TWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQ"
                    }
                }
            },
            "ImageResource": {
                "title": "Image",
                "description": "Card's image",
                "properties": {
                    "id": {
                        "description": "Image's ID",
                        "type": "integer",
                        "format": ""
                    },
                    "name": {
                        "description": "Fila name",
                        "type": "string",
                        "format": ""
                    },
                    "extension": {
                        "description": "File's extension",
                        "type": "string",
                        "format": ""
                    },
                    "file": {
                        "description": "Base 64 code",
                        "type": "string",
                        "format": ""
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Dark Margician",
                    "extension": "jpg",
                    "file": "TWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQ"
                }
            },
            "SubtypeResource": {
                "title": "Subtype",
                "description": "Card's subtype",
                "properties": {
                    "id": {
                        "description": "Subtype's ID",
                        "type": "integer",
                        "format": ""
                    },
                    "name": {
                        "description": "Subtype's name",
                        "type": "string",
                        "format": ""
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Normal Monster"
                }
            }
        }
    }
}